Carry arg_plus at 64 bits through txn_append_arg_patch
The shared fold helper was written against a 32-bit arg_plus; main widened
address_patch's arg_plus to its real 64-bit width in #3718. Merging the two
left the helper truncating every offset above 4 GiB -- and worse, the fold
itself wrapped: 0xFFFFFFFF + 0x80000000 lost its carry instead of landing in
the high word. Caught by Targets/NPU/npu_address_patch_arg_plus_64bit.mlir.

+++ include/aie/Runtime/TxnEncoding.h
@@ -259,7 +259,7 @@ inline void txn_append_address_patch(std::vector<uint32_t> &txn, uint32_t addr,
 // when the target runtime needs it. Both the static binary emitter and the
 // generated C++ builder route through here, so the fold rule has one body.
 inline void txn_append_arg_patch(std::vector<uint32_t> &txn, uint32_t addr,
-                                 int32_t arg_idx, uint32_t arg_plus,
+                                 int32_t arg_idx, uint64_t arg_plus,
                                  bool fold_ddr_addr_offset) {
   if (fold_ddr_addr_offset &&
       static_cast<uint32_t>(arg_idx) >= kNumFirmwareTranslatedArgs)
